requirements.txt
anyio==4.9.0
arabic-reshaper==3.0.0
certifi==2025.4.26
chardet==5.2.0
defusedxml==0.7.1
fonttools==4.58.0
fpdf2==2.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
nest-asyncio==1.6.0
pillow==11.2.1
python-bidi==0.6.6
python-telegram-bot[rate-limiter]==22.1
reportlab==4.4.1
sniffio==1.3.1
fastapi==0.110.2
uvicorn==0.29.0
